+++ server/middleware/validateDirQuery.js
@@ -17,10 +17,14 @@ module.exports = async function checkQuery(req, res, next) {
     !fileTypes.split(',').every(type => SUPPORTED_TYPES.includes(type))
   ) {
     return res.status(400).send({
-      msg: `fileTypes query param must match given types: ${String(
+      msg: `fileTypes query param must match one of types: ${String(
         SUPPORTED_TYPES
       )}`,
     })
+  }
+
+  if (path.toLowerCase() === 'home') {
+    return next()
   }
 
   try {
@@ -28,9 +32,8 @@ module.exports = async function checkQuery(req, res, next) {
     if (!lstat.isDirectory()) {
       throw new Error('Not a directory')
     }
+    return next()
   } catch {
     return res.status(400).send({ msg: `"${path}" is not a valid directory` })
   }
-
-  next()
 }
