+++ server/package.json
@@ -3,24 +3,26 @@
   "version": "1.0.0",
   "description": "",
   "main": "electron-starter.js",
+  "license": "MIT",
   "scripts": {
     "start": "concurrently \"yarn run server:prod\" \"cross-env BROWSER=chrome yarn run start-client\"",
     "server:dev": "cross-env NODE_ENV=development WHITE_LIST=http://localhost:3003 nodemon src/index.js",
     "server:prod": "cross-env NODE_ENV=production WHITE_LIST=http://localhost:3003 node src/index.js",
     "start-client": "cd .. && cd client && cross-env HOST=localhost PORT=3003 yarn start",
+    "build-client": "cd .. && cd client && yarn run build",
     "start-electron": "wait-on http://localhost:3003 && cross-env NODE_ENV=production WHITE_LIST=http://localhost:3003 electron electron-starter",
     "electron-dev": "concurrently \"cross-env BROWSER=none yarn run start-client\" \"yarn run start-electron\"",
+    "electron-build": "electron-builder",
+    "build-electron": "concurrently \"yarn run build-client\" \"yarn run electron-build\"",
     "test": "jest --watchAll=true --env=node",
     "coverage": "jest --coverage --watchAll=false --env=node"
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
   "dependencies": {
     "concurrently": "^5.3.0",
     "cors": "^2.8.5",
     "cross-env": "^7.0.2",
-    "electron-builder": "^22.9.1",
     "electron-is-dev": "^1.2.0",
     "express": "^4.17.1",
     "morgan": "^1.10.0",
@@ -29,6 +31,7 @@
   },
   "devDependencies": {
     "electron": "^11.0.2",
+    "electron-builder": "^22.9.1",
     "jest": "24.9.0",
     "nodemon": "^2.0.4"
   },
